requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=3.14
psycopg2-binary>=2.9.9
Pillow>=9.0
SpeechRecognition
pydub
opencv-python-headless
langchain
langchain-core
langchain-community
langchain-groq
python-dotenv
llama-index
llama-index-llms-groq
fastembed
llama-index-embeddings-fastembed
requests
gunicorn
whitenoise
dj-database-url
django-anymail[sendgrid]
django-import-export
django-cloudinary-storage
cloudinary
